# packages/sandbox/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Open Browser Sandbox API",
    "description": "API documentation for Open Browser Sandbox service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/sdk/config": {
      "get": {
        "operationId": "getSdkConfig",
        "description": "Get SDK Config",
        "responses": {
          "200": {
            "description": "SDK config retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/sdk/agent": {
      "get": {
        "operationId": "getSdkAgent",
        "description": "Get SDK Agent",
        "responses": {
          "200": {
            "description": "SDK agent retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/config": {
      "get": {
        "operationId": "getConfig",
        "description": "Get Current Instance Config",
        "responses": {
          "200": {
            "description": "Config retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sdkType": {
                      "type": "string",
                      "enum": [
                        "OPENCODE",
                        "CLAUDE_CODE"
                      ]
                    },
                    "directory": {
                      "type": "string"
                    },
                    "initialized": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "sdkType",
                    "directory",
                    "initialized"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}